requirements.txt
fastapi
uvicorn
debugpy
httpx
beautifulsoup4
python-dotenv
prometheus-fastapi-instrumentator
langchain
pydantic
opencv-python

# recomendation engine
lancedb
pandas
pytest
numpy
colorama